requirements.txt
tensorflow==2.4.1
keras==2.4.2
pandas
numpy
pyYaml
efficientnet
